+++ open_py/config.py
@@ -1,0 +1,5 @@
+api_prefix = 'api/'
+remote_api = ''
+
+# base_url = 'https://api.open-platform.zensoft.io'
+base_url = 'https://api.openfuture.io'