+++ openplatform/config.py
@@ -1,5 +1,0 @@
-api_prefix = 'api/'
-remote_api = ''
-
-# base_URL= 'https://api.openfuture.io for discussion
-base_url = 'https://api.open-platform.zensoft.io'