+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2018-08-02
+### Changed
+- Renamed module to open_py
+- Renamed class to OpenPy
+
+
 ## [1.0.1] - 2018-08-01
 ### Changed
 - Update the README to match the new way of importing the package
@@ -27,5 +33,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Get all transactions API
 - Tests
 
+[2.0.0]: https://github.com/OpenFuturePlatform/open-api-python-sdk/compare/v2.0.0...v1.0.1
+[1.0.1]: https://github.com/OpenFuturePlatform/open-api-python-sdk/compare/v1.0.1...v1.0.0
 [1.0.1]: https://github.com/OpenFuturePlatform/open-api-python-sdk/compare/v1.0.1...v1.0.0
 [1.0.0]: https://github.com/OpenFuturePlatform/open-api-python-sdk/compare/e8d583f41e071d56dbc6dcf924bcc3a35c85b302...v1.0.0